moves the functionality into a fully fledged report and adds a report list page
