+++ opat/templates/reports/opat/nors_report.html
@@ -1,0 +1,116 @@
+{% extends "reporting/report_detail.html" %}
+{% load forms %}
+{% block report_button %}
+  <div class="infectious-diseases-report-button">
+    <span ng-show="!report.criteria.quarter">
+      <h4>Please choose a quarter.</h4>
+    </span>
+    <span ng-show="report.criteria.quarter">
+      {{ block.super }}
+    </span>
+  </div>
+{% endblock %}
+
+
+{% block report_choices %}
+<div class="row ">
+ {% for row in object.reports_available %}
+  <div class="col-sm-6 col-md-3 content-offset-below-25">
+    <label ng-change="report.refresh()" class="btn btn-default infectious-diseases-report-spacing" ng-model="report.criteria.quarter" uib-btn-radio="'{{ row.value }}'">
+      <h3>{{ row.display_name }}</h3>
+      <hr />
+      <div class="row content-offset-10 content-offset-below-10">
+        <div class="col-md-12">
+          <div class="row content-offset-below-10 text-left">
+            <div class="col-xs-10">
+              Total Episodes:
+            </div>
+            <div class="col-xs-2">
+              {{ row.report.episodes }}
+            </div>
+          </div>
+          <div class="row content-offset-below-10 text-left">
+            <div class="col-xs-10">
+              Total Treatment Days Saved:
+            </div>
+            <div class="col-xs-2">
+              {{ row.report.total_treatment_days_saved }}
+            </div>
+          </div>
+          <div class="row content-offset-below-10 text-left">
+            <div class="col-xs-10">
+              Total Line Events:
+            </div>
+            <div class="col-xs-2">
+              {{ row.report.total_line_events }}
+            </div>
+          </div>
+          <div class="row content-offset-below-10 text-left">
+            <div class="col-xs-10">
+              Total Drug Events:
+            </div>
+            <div class="col-xs-2">
+              {{ row.report.total_drug_events }}
+            </div>
+          </div>
+        </div>
+      </div>
+    </label>
+  </div>
+  {% comment %}
+  <div class="col-sm-6 col-md-3">
+    <div class="panel panel-default">
+      <div class="panel-heading">
+        <h3>{{ row.display_name }}</h3>
+      </div>
+      <div class="panel-body">
+        <div class="row content-offset-10 content-offset-below-10">
+          <div class="col-md-12">
+            <div class="row content-offset-below-10">
+              <div class="col-md-10">
+                Total Episodes:
+              </div>
+              <div class="col-md-2">
+                {{ row.report.episodes }}
+              </div>
+            </div>
+            <div class="row content-offset-below-10">
+              <div class="col-md-10">
+                Total Treatment Days Saved:
+              </div>
+              <div class="col-md-2">
+                {{ row.report.total_treatment_days_saved }}
+              </div>
+            </div>
+            <div class="row content-offset-below-10">
+              <div class="col-md-10">
+                Total Line Events:
+              </div>
+              <div class="col-md-2">
+                {{ row.report.total_line_events }}
+              </div>
+            </div>
+            <div class="row content-offset-below-10">
+              <div class="col-md-10">
+                Total Drug Events:
+              </div>
+              <div class="col-md-2">
+                {{ row.report.total_drug_events }}
+              </div>
+            </div>
+          </div>
+        </div>
+        <div class="row">
+          <div class="col-md-12">
+            <button class="btn btn-primary btn-lg">
+              <i class="fa fa-cloud-download"></i> Get Report
+            </button>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+  {% endcomment %}
+ {% endfor %}
+</div>
+{% endblock %}
